% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{surveyfreq}
\alias{surveyfreq}
\title{Generate the weighted frequency table}
\usage{
surveyfreq(arr, based_wt = NULL, rep_wts = NULL, includeNA = FALSE,
  NA_wt = 1, varmethod = "jackknife", BRRfay = 0)
}
\arguments{
\item{arr}{A \code{vector} (integer, double, character, logical or complex) of dimension M}

\item{based_wt}{A \code{vector} of dimension of M}

\item{rep_wts}{A \code{matrix} of M x N}

\item{includeNA}{Logical. Should missing values (including NaN) be removed?}

\item{NA_wt}{double Assign a constance value if weight is missing}

\item{varmethod}{Character, "jackknife" or "BRR"}

\item{BRRfay}{0 - 1, only available when varmethod=="BRR"}
}
\description{
Generate the weighted frequency table with jackknife or BRR similar with PROC SURVEYFREQ procedure.
It takes roughly 10+ seconds to process 10
}
\examples{
\dontrun{
# Generate a vector, based weight and replicated weights
arr=sample(1:10, 1000, replace=TRUE)
based_wt=runif(1000)
rep_wts=matrix(runif(1000*80), ncol=80)

# Generate table
surveyfreq2(arr, based_wt, rep_wts)
}
}
\author{
Wenke Huang
}
